requirements.txt
attrs==25.3.0
blinker==1.9.0
capmonster-python==2.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.2.1
Flask==3.0.0
googlemaps==4.10.0
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.31.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stripe==12.2.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==1.26.14
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
